chore: Add zoxide to zsh plugins
Added zoxide to the list of zsh plugins in the .zsh_plugins.txt file. This plugin provides fast directory navigation by allowing users to jump to previously visited directories with just a few keystrokes. The 'kind:defer' option ensures that zoxide is loaded lazily, improving the startup time of the shell. This addition will enhance the user experience and productivity when navigating directories in the terminal.